% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scen_data.R
\name{generate_scen_data}
\alias{generate_scen_data}
\title{Generate simulation data from pre-specified scenarios}
\usage{
generate_scen_data(scen, include_truth = TRUE)
}
\arguments{
\item{scen}{A row from the scen_param data set}

\item{include_truth}{Boolean, will the true treatment effect be included in the outcome data-set?}
}
\value{
A data frame
}
\description{
Generate simulation data from simulation scenarios used in XYZ. The
scenarios were derived so that the prognostic part of the model on
control as an R^2 of 0.32 for continuous data and an AUC of 0.66
for binary data. The coefficient b0 was calculated in each scenario so
that the overall test for a treatment effect has a power of 50% for
2.5% one-sided. There are four main scenarios (defined by the
expressions for prognostic and predictive part) replicated across
continuous and binary endpoints. Within each scenario there are 5
sub-scenarios corresponding to different selections of b1. The
third scenario in each of the sub-scenarios correspond to the scenario
where the interaction test (under the true model) has 80% power
(for 20% two-sided, or 10% one-sided as only b1 > 0 is
considered). The other 4 scenarios correspond to 0, 0.5, 1.5, 2 times
the b1 value that provides 80% power of the interaction test. In
total there are hence 4x2x5=40 scenarios.
}
\examples{
data(scen_param) ## scenarios used in XYZ
dat <- generate_scen_data(scen = scen_param[1, ])
}
