% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_X_syn.R
\name{generate_X_syn}
\alias{generate_X_syn}
\title{Generate X from data "syn"}
\usage{
generate_X_syn(n)
}
\arguments{
\item{n}{Number of observations}
}
\value{
A data frame of predictors
}
\description{
Generate data from synthetic clinical trial data with dimension n
and 30 variables.
}
\details{
The data are generated based on the data-set data_syn which is not
exported from the package namespace. This data-set was generated
from pooled set of clinical trial data using the synthpop function
from the synthpop R package. data_syn has 1934 rows and 30
variables. Variable names have been removed. Continuous variables
have been scaled to the interval [0,1], levels of the categorical
variable have been blinded. For information on the synthetic data
which synthpop generates from, see the vignette: Description of
included synthetic data.
}
\examples{
X <- generate_X_syn(n=100)

}
