% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_y.R
\name{generate_y}
\alias{generate_y}
\title{Generate outcome variable y and merge with predictor data matrix X.}
\usage{
generate_y(
  X,
  trt,
  prog,
  pred,
  b0,
  b1 = NULL,
  sd_te = NULL,
  type = c("binary", "continuous", "count", "survival"),
  sigma_error = 1,
  theta = 1,
  cens_time = NULL,
  lambda0 = NULL,
  include_truth = FALSE,
  sign_better = 1
)
}
\arguments{
\item{X}{matrix/dataframe of predictor variables}

\item{trt}{Binary treatment indicator variable}

\item{prog}{Character variable giving expression for prognostic
effects (defined in terms of names in the X matrix)}

\item{pred}{Character variable giving expression for predictive
effects (defined in terms of names in the X matrix)}

\item{b0}{Treatment (main effect)}

\item{b1}{Coefficient of the predictive effects defined in pred}

\item{sd_te}{Standard deviation of the treatment effects defined
via pred. If given b1 is ignored. For binary data this is assumed
to be on the log-odds ratio scale. For survival and count data
this is assumed to be on the log scale.}

\item{type}{Outcome data type to generate ("continuous", "binary",
"count" and "survival" are allowed here)}

\item{sigma_error}{Residual error, only needed for type =
"continuous"}

\item{theta}{Overdispersion parameter, only needed for data_type =
"count" (variance of neg bin in this parameterization is mu +
mu^2/theta)}

\item{cens_time}{Function to generate the censoring time, only
needed for data_type = "survival"}

\item{lambda0}{Intercept of exponential regression (on non-log scale)}

\item{include_truth}{boolean, will the true prognostic and
predictive effects be included in the outcome data-set?}

\item{sign_better}{whether larger response is better (used to
determine whether b1 is negative or positive if not given)}
}
\value{
Data set
}
\description{
Generate simulated data according to the model f(X) = f_prog(X) +
trt*(b0 + b1*f_pred(X)) for different outcome distributions.  For
continuous data f(X) is the conditional mean, for binary data the
logit response probility, for count data the log-rate and for
survival data the log-hazard rate.
}
\examples{
X <- generate_X_dist(n = 10000, p = 10, rho = 0.5)
## observational data set
trt <- generate_trt(n = nrow(X), type = "random", X = X, prop = "X2")
dat <- generate_y(X, trt,
  prog = "0.5*((X1=='Y')+X3)",
  pred = "X3>0", b0 = 0, b1 = 1,
  type = "continuous", sigma_error = 3
)


#### generate data from user specified covariate X
X <- sapply(1:10, function(ii) {
  rnorm(500)
})
X <- as.data.frame(X)
colnames(X) <- paste0("Z", 1:10)
trt <- generate_trt(nrow(X), p_trt = 0.5)
dat <- generate_y(X, trt,
  prog = "0.5*Z1",
  pred = "(Z5>0)", b0 = 0, b1 = 1,
  type = "binary"
)
glm(Y ~ trt * ., data = dat, family = binomial)
}
