% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scen_param.R
\docType{data}
\name{scen_param}
\alias{scen_param}
\title{Pre-specified simulation scenarios from XYZ}
\format{
See more details on the choice of parameters considered in
\url{https://htmlpreview.github.io/?https://raw.githubusercontent.com/Sophie-Sun/benchtm/master/docs/articles/a01_User_guidance.html}
}
\usage{
scen_param
}
\description{
The scenarios were derived so that the prognostic part of the model
on the control arm has an R^2 of 0.32 for continuous data and an
AUC of 0.66 for binary data. The coefficient b0 was calculated in
each case so that the overall test for a treatment effect has a
power of 50% for 2.5% one-sided. There are four main scenarios
(defined by the expressions for prognostic and predictive part)
replicated across continuous and binary endpoints. Within each
scenario there are 5 sub-scenarios corresponding to different
selections of b1. The third case in each of the sub-scenarios
correspond to the case where the interaction test (under the true
model) has 80% power (for 20% two-sided, or 10% one-sided as only
b1 > 0 is considered). The other 4 cases correspond to 0, 0.5, 1.5,
2 times the b1 value that provides 80% power of the interaction
test. The column b1_rel gives the size of b1 relative to the b1
value that provides 80% power. In total there are 4x2x5=40
scenarios.
}
\keyword{datasets}
